Map generator with GUI.
Added terrain generator (water/terrain). TerraGUI is a test and generator at the same time. Map is 10x10. generate_terrain.py has to be rewritten as a class.

StreetGenerator_v2 with changed way road is created.

City has to be bigger than 6x6 so that code can generate map correctly (infinite loop). This has to be corrected in next version